+++ packages/crawler/src/site.ts
@@ -18,6 +18,13 @@ function parseSitemapUrl(sitemapItem: SitemapItem) {
 export const crawlSite = async ({ url, pathname, interval = 0 }: Site) => {
   logger.info(`Start crawl from sitemap ${url}`, { pathname });
 
+  const key = `${url}-${pathname}`;
+
+  if (crawlBlockletRunningMap.has(key)) {
+    logger.info(`Crawl from sitemap ${url} ${pathname} is already running, skip`);
+    return [];
+  }
+
   const sitemapList = await getSitemapList(url);
   const pathnameRegex = new RegExp(pathname);
 
@@ -29,9 +36,15 @@ export const crawlSite = async ({ url, pathname, interval = 0 }: Site) => {
 
   logger.info(`Found ${sitemapItems.length} sitemap items which match ${pathname} from ${url}`);
 
-  const crawlableItems = (
-    await Promise.all(
-      sitemapItems.map(async ({ url, sitemapItem }) => {
+  let processCount = 0;
+  crawlBlockletRunningMap.set(key, true);
+
+  try {
+    const jobIds = await pMap(
+      sitemapItems,
+      async ({ url, sitemapItem }) => {
+        processCount++;
+
         const snapshot = await Snapshot.findOne({ where: { url: formatUrl(url) } });
 
         if (snapshot?.lastModified) {
@@ -48,20 +61,8 @@ export const crawlSite = async ({ url, pathname, interval = 0 }: Site) => {
           }
         }
 
-        return { url, sitemapItem };
-      }),
-    )
-  ).filter(Boolean) as { url: string; sitemapItem: SitemapItem }[];
+        logger.debug(`Sitemap process ${processCount} / ${sitemapItems.length}`);
 
-  logger.info(`Found ${crawlableItems.length} pages to crawl from sitemap ${url}`, { pathname });
-
-  const key = `${url}-${pathname}`;
-  crawlBlockletRunningMap.set(key, crawlableItems);
-
-  try {
-    const jobIds = await pMap(
-      crawlableItems,
-      ({ url, sitemapItem }) => {
         return crawlUrl({
           url,
           lastModified: sitemapItem.lastmod,
@@ -69,8 +70,9 @@ export const crawlSite = async ({ url, pathname, interval = 0 }: Site) => {
           includeHtml: true,
         });
       },
-      { concurrency: config.siteCron.concurrency },
+      { concurrency: config.siteCron.sitemapConcurrency },
     );
+
     return jobIds;
   } catch (error) {
     logger.error(`Failed to crawl from sitemap ${url} ${pathname}`, error);